코멘트 하는중

+++ src/main/java/site/metacoding/white/service/BoardService.java
@@ -12,9 +12,11 @@ import site.metacoding.white.domain.Board;
 import site.metacoding.white.domain.BoardRepository;
 import site.metacoding.white.domain.UserRepository;
 import site.metacoding.white.dto.BoardReqDto.BoardSaveReqDto;
+import site.metacoding.white.dto.BoardReqDto.BoardUpdateReqDto;
 import site.metacoding.white.dto.BoardRespDto.BoardAllRespDto;
 import site.metacoding.white.dto.BoardRespDto.BoardDetailRespDto;
 import site.metacoding.white.dto.BoardRespDto.BoardSaveRespDto;
+import site.metacoding.white.dto.BoardRespDto.BoardUpdateRespDto;
 
 // 트랜잭션 관리
 // Dto 변환해서 컨트롤러에게 돌려줘야 함
@@ -49,12 +51,15 @@ public class BoardService {
     }
 
     @Transactional
-    public void update(Long id, Board board) {
+    public BoardUpdateRespDto update(BoardUpdateReqDto boardUpdateReqDto) {
+        Long id = boardUpdateReqDto.getId();
         Optional<Board> boardOP = boardRepository.findById(id);
         if (boardOP.isEmpty()) {
             throw new RuntimeException("해당 " + id + "로 수정을 할 수 없습니다");
         }
-        boardOP.get().update(board.getTitle(), board.getContent());
+        Board boardPS = boardOP.get();
+        boardPS.update(boardUpdateReqDto.getTitle(), boardUpdateReqDto.getContent());
+        return new BoardUpdateRespDto(boardPS);
     } // 트랜잭션 종료시 => 더티체킹을 함
 
     @Transactional(readOnly = true)
@@ -65,12 +70,17 @@ public class BoardService {
         for (Board board : boardList) {
             boardAllRespDtoList.add(new BoardAllRespDto(board));
         }
-
         return boardAllRespDtoList;
     }
 
     @Transactional
     public void deleteById(Long id) {
-        boardRepository.deleteById(id);
+        Optional<Board> boardOP = boardRepository.findById(id);
+        if (boardOP.isPresent()) {
+            boardRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("해당 " + id + "로 삭제를 할 수 없습니다");
+        }
+
     }
 }
